Use a network-helper mod to remove duplicates
