+++ src/apple/network.rs
@@ -6,8 +6,8 @@ use std::collections::{hash_map, HashMap};
 use std::net::Ipv4Addr;
 use std::ptr::null_mut;
 
-use crate::common::{InterfaceAddress, MacAddr};
-use crate::network::get_interface_address;
+use crate::common::MacAddr;
+use crate::network::refresh_networks_addresses;
 use crate::{NetworkExt, NetworksExt, NetworksIter};
 
 macro_rules! old_and_new {
@@ -156,23 +156,6 @@ impl Networks {
                 }
             }
         }
-
-        if let Ok(iter) = get_interface_address() {
-            for (name, ifa) in iter {
-                if let Some(interface) = self.interfaces.get_mut(&name) {
-                    match ifa {
-                        InterfaceAddress::MAC(mac_addr) => {
-                            interface.mac_addr = mac_addr;
-                        }
-                        InterfaceAddress::IPv4(addr, mask) => {
-                            interface.ipv4_addr = addr;
-                            interface.ipv4_mask = mask;
-                        }
-                        _ => {}
-                    }
-                }
-            }
-        }
     }
 }
 
@@ -188,6 +171,7 @@ impl NetworksExt for Networks {
         }
         self.update_networks();
         self.interfaces.retain(|_, data| data.updated);
+        refresh_networks_addresses(&mut self.interfaces);
     }
 
     fn refresh(&mut self) {
@@ -211,9 +195,12 @@ pub struct NetworkData {
     errors_out: u64,
     old_errors_out: u64,
     updated: bool,
-    mac_addr: MacAddr,
-    ipv4_addr: Ipv4Addr,
-    ipv4_mask: Ipv4Addr,
+    /// MAC address
+    pub(crate) mac_addr: MacAddr,
+    /// IPv4 address
+    pub(crate) ipv4_addr: Ipv4Addr,
+    /// IPv4 subnet mask
+    pub(crate) ipv4_mask: Ipv4Addr,
 }
 
 impl NetworkExt for NetworkData {
