+++ src/lib.rs
@@ -25,7 +25,8 @@ cfg_if::cfg_if! {
         mod apple;
         use apple as sys;
         extern crate core_foundation_sys;
-        pub(crate) mod network;
+        mod network_helper_nix;
+        use network_helper_nix as network_helper;
 
         #[cfg(test)]
         pub(crate) const MIN_USERS: usize = 1;
@@ -34,6 +35,8 @@ cfg_if::cfg_if! {
         use windows as sys;
         extern crate winapi;
         extern crate ntapi;
+        mod network_helper_win;
+        use network_helper_win as network_helper;
 
         #[cfg(test)]
         pub(crate) const MIN_USERS: usize = 1;
@@ -41,7 +44,8 @@ cfg_if::cfg_if! {
         mod linux;
         use linux as sys;
         pub(crate) mod users;
-        pub(crate) mod network;
+        mod network_helper_nix;
+        use network_helper_nix as network_helper;
 
         #[cfg(test)]
         pub(crate) const MIN_USERS: usize = 1;
@@ -49,7 +53,8 @@ cfg_if::cfg_if! {
         mod freebsd;
         use freebsd as sys;
         pub(crate) mod users;
-        pub(crate) mod network;
+        mod network_helper_nix;
+        use network_helper_nix as network_helper;
 
         #[cfg(test)]
         pub(crate) const MIN_USERS: usize = 1;
@@ -78,6 +83,7 @@ pub use c_interface::*;
 mod c_interface;
 mod common;
 mod debug;
+mod network;
 mod system;
 mod traits;
 mod utils;
